git-svn-id: http://svn.ilias.de/svn/ilias/trunk@19249 21b2c9ec-7c21-0410-8b45-9bfb8ed2bfc5

+++ Modules/Test/classes/class.ilTestExport.php
@@ -503,19 +503,23 @@ class ilTestExport
 		// test participant result export
 		foreach ($data->getParticipants() as $active_id => $userdata) 
 		{
-			$resultsheet =& $workbook->addWorksheet(ilExcelUtils::_convert_text($userdata->getName()));
-			$pass = $userdata->getScoredPass();
-			$row = 0;
-			$resultsheet->writeString($row, 0, ilExcelUtils::_convert_text(sprintf($this->lng->txt("tst_result_user_name_pass"), $pass+1, $userdata->getName())), $format_bold);
-			$row += 2;
-			if (is_object($userdata) && is_array($userdata->getQuestions($pass)))
+			$username = (!is_null($userdata) && ilExcelUtils::_convert_text($userdata->getName())) ? ilExcelUtils::_convert_text($userdata->getName()) : "ID $active_id";
+			$resultsheet =& $workbook->addWorksheet($username);
+			if (method_exists($resultsheet, "writeString"))
 			{
-				foreach ($userdata->getQuestions($pass) as $question)
-				{ 
-					$question =& $this->test_obj->_instanciateQuestion($question["aid"]);
-					if (is_object($question))
-					{
-						$row = $question->setExportDetailsXLS($resultsheet, $row, $active_id, $pass, $format_title, $format_bold);
+				$pass = $userdata->getScoredPass();
+				$row = 0;
+				$resultsheet->writeString($row, 0, ilExcelUtils::_convert_text(sprintf($this->lng->txt("tst_result_user_name_pass"), $pass+1, $userdata->getName())), $format_bold);
+				$row += 2;
+				if (is_object($userdata) && is_array($userdata->getQuestions($pass)))
+				{
+					foreach ($userdata->getQuestions($pass) as $question)
+					{ 
+						$question =& $this->test_obj->_instanciateQuestion($question["aid"]);
+						if (is_object($question))
+						{
+							$row = $question->setExportDetailsXLS($resultsheet, $row, $active_id, $pass, $format_title, $format_bold);
+						}
 					}
 				}
 			}
